refactor: (#513) Report 도메인 코드 가독성 리팩토링
